windows/requirements.txt
paho-mqtt==2.1.0
Pillow==11.1.0
opencv-python==4.11.0.86
pyautogui==0.9.54
pyperclip==1.9.0
nanoid==2.0.0
discord==2.3.2
python-dotenv==1.0.1